Refactor folder structure, add a 'pubsub' path
Added /post/ to path for PubSub messages to break up server by sections.

Added test script that uses jest, part of completing testing setup.

I moved GameLocator to a utils folder to keep the space more organised.
