+++ rest-client/basic_req.rest
@@ -1,4 +1,4 @@
-POST http://localhost:8080/ HTTP/1.1
+POST http://localhost:8080/pubsub/ HTTP/1.1
 content-type: application/json
 
 {
